Remove error message on persons with multiple affiliations (#20)
eg. it errored with:

```
"affiliation" must be a Organization object or URI, not [{"@type":"Organization","@id":"http://example.org/"},{"@type":"Organization","@id":"http://example.com/"}]'
```

which is not true.

This will cause the field to be populated with
`[object Object],[object Object]` instead of the actual values, which we
need to fix eventually, but at least it doesn't incorrectly imply arrays
are forbidden by Codemeta.